refactor: 도메인 패키지 구조를 서브패키지 컨벤션으로 통일
- util/validation 중복 디렉토리 평면화
- developer 도메인 controller/·service/ 분리
- 전 도메인 models.py → model/<도메인>_models.py
