+++ app/domains/agent/service/agent_service.py
@@ -3,9 +3,9 @@ from typing import Any, Dict, List, Optional
 
 from app.common.security import utc_now
 from app.common.logging.logging_config import get_logger
-from app.domains.agent.models import AgentDecision, ChatMessage, ChatSession
+from app.domains.agent.model.agent_models import AgentDecision, ChatMessage, ChatSession
 from app.domains.agent.repository.chat_session_repository import ChatSessionRepository
-from app.domains.progress.models import StudentWeaknessProfile
+from app.domains.progress.model.progress_models import StudentWeaknessProfile
 from app.domains.progress.service.learning_record_service import LearningRecordService
 from app.infrastructure.db.vector.vector_db import get_vector_db
 from app.infrastructure.embedding.embedding_model import get_embedding_model
