+++ app/domains/classroom/service/classroom_service.py
@@ -1,8 +1,8 @@
 import secrets
 from datetime import datetime, timezone
 
-from app.domains.auth.models import User
-from app.domains.classroom.models import Classroom
+from app.domains.auth.model.auth_models import User
+from app.domains.classroom.model.classroom_models import Classroom
 from app.domains.classroom.repository.classroom_repository import ClassroomRepository
