+++ app/domains/content/stage3/router.py
@@ -6,7 +6,7 @@ from app.domains.content.stage3.schemas import (
     Stage3ProgressResponse,
 )
 from app.domains.auth.dependency.auth_dependencies import get_current_user
-from app.domains.auth.models import User
+from app.domains.auth.model.auth_models import User
 from app.domains.content.stage3.service import (
     DEFAULT_STAGE3_LESSON_ID,
     get_stage3_service,
