+++ app/domains/instruction/service/generation.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, Field
 
-from app.domains.instruction.models import GeneratedProblem
+from app.domains.instruction.model.instruction_models import GeneratedProblem
 from app.domains.progress.util import CONCEPT_KEY_BY_ANSWER
 from app.infrastructure.external.openai_client import OpenAIClient
