+++ app/domains/progress/controller/teacher_student_router.py
@@ -10,7 +10,7 @@ progress 도메인에 위치한 이유:
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 
 from app.domains.auth.dependency.auth_dependencies import get_current_teacher
-from app.domains.auth.models import User
+from app.domains.auth.model.auth_models import User
 from app.domains.classroom.dependency.classroom_dependencies import get_classroom_service
 from app.domains.classroom.service.classroom_service import ClassroomService
 from app.domains.progress.dependency.learning_record_dependencies import get_learning_record_service
