+++ tests/agent/test_session_service.py
@@ -4,10 +4,10 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from app.domains.agent.models import AgentDecision, ChatMessage, ChatSession
+from app.domains.agent.model.agent_models import AgentDecision, ChatMessage, ChatSession
 from app.domains.agent.repository.chat_session_repository import ChatSessionRepository
 from app.domains.agent.service.agent_service import AgentService, ChatSessionService
-from app.domains.progress.models import StudentWeaknessProfile, WeakConcept
+from app.domains.progress.model.progress_models import StudentWeaknessProfile, WeakConcept
 
 
 # ---------------------------------------------------------------------------
