+++ tests/classroom/test_classroom_service.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from app.domains.auth.models import User
+from app.domains.auth.model.auth_models import User
 from app.domains.classroom.service.classroom_service import ClassroomService
