+++ tests/classroom/test_teacher_controllers.py
@@ -4,7 +4,7 @@ from types import SimpleNamespace
 import pytest
 from fastapi import HTTPException
 
-from app.domains.auth.models import User
+from app.domains.auth.model.auth_models import User
 from app.domains.classroom.controller.teacher_class_router import list_class_students, list_my_classes
 from app.domains.progress.controller.teacher_student_router import get_student_records
