+++ tests/structure/test_domain_structure.py
@@ -9,7 +9,7 @@ def test_domain_packages_are_importable():
         "app.domains.developer.service.indexing_service",
         "app.domains.developer.controller.admin_router",
         "app.domains.auth",
-        "app.domains.auth.models",
+        "app.domains.auth.model.auth_models",
         "app.domains.auth.repository.auth_repository",
         "app.domains.auth.controller.auth_router",
         "app.domains.auth.schema.auth_schemas",
@@ -18,7 +18,7 @@ def test_domain_packages_are_importable():
         "app.domains.classroom.controller.student_class_router",
         "app.domains.progress.controller.teacher_student_router",
         "app.domains.classroom.dependency.classroom_dependencies",
-        "app.domains.classroom.models",
+        "app.domains.classroom.model.classroom_models",
         "app.domains.classroom.repository.classroom_repository",
         "app.domains.classroom.service.classroom_service",
         "app.domains.classroom.schema.classroom_schemas",
@@ -28,7 +28,7 @@ def test_domain_packages_are_importable():
         "app.domains.instruction.controller.teacher_router",
         "app.domains.instruction.service.generation",
         "app.domains.instruction.validation",
-        "app.domains.instruction.models",
+        "app.domains.instruction.model.instruction_models",
         "app.domains.instruction.repository.assignment_repository",
         "app.domains.instruction.schema.instruction_schemas",
         "app.domains.instruction.service.instruction_service",
@@ -42,11 +42,11 @@ def test_domain_packages_are_importable():
         "app.domains.progress.controller.student_progress_router",
         "app.domains.content.stage3.router",
         "app.domains.content",
-        "app.domains.content.models",
+        "app.domains.content.model.content_models",
         "app.domains.content.repository.catalog_repository",
         "app.domains.content.schema.catalog_schemas",
         "app.domains.content.service.catalog_service",
-        "app.domains.progress.models",
+        "app.domains.progress.model.progress_models",
         "app.domains.progress.repository.learning_record_repository",
         "app.domains.progress.schema.learning_record_schemas",
         "app.domains.progress.service.learning_record_service",
